+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irFunctionAccessExpression.kt
@@ -1,7 +1,12 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package com.javiersc.kotlin.compiler.extensions.ir
 
 import com.javiersc.kotlin.compiler.extensions.common.fqName
+import org.jetbrains.kotlin.ir.declarations.IrParameterKind
+import org.jetbrains.kotlin.ir.declarations.IrValueParameter
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
+import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
 import org.jetbrains.kotlin.ir.util.hasAnnotation
 import org.jetbrains.kotlin.ir.util.kotlinFqName
@@ -26,3 +31,13 @@ public inline fun <reified T : Annotation> IrFunctionAccessExpression.hasAnnotat
 
 public inline fun IrFunctionAccessExpression.hasAnnotation(annotation: FqName): Boolean =
     annotations.hasAnnotation(annotation)
+
+public val IrFunctionAccessExpression.extensionReceiver: IrExpression?
+    get() = getExtensionReceiverOrNull()
+
+internal inline fun IrFunctionAccessExpression.getExtensionReceiverOrNull(): IrExpression? {
+    val parameters: List<IrValueParameter> = symbol.owner.parameters
+    val extensionReceiverIndex: Int =
+        parameters.indexOfFirst { it.kind == IrParameterKind.ExtensionReceiver }
+    return arguments[extensionReceiverIndex]
+}